Factorize some front views behavior (el, )
